+++ connect/api/v2/template_projects/views.py
@@ -1,7 +1,8 @@
 from rest_framework.viewsets import ModelViewSet
 from rest_framework.response import Response
-from connect.template_projects.models import TemplateType, TemplateFeature, TemplateAI
-from .serializers import TemplateTypeSerializer, RetrieveTemplateSerializer, TemplateFeatureSerializer, TemplateAISerializer
+from rest_framework.permissions import IsAuthenticated
+from connect.template_projects.models import TemplateType, TemplateFeature, TemplateSuggestion
+from .serializers import TemplateTypeSerializer, RetrieveTemplateSerializer, TemplateFeatureSerializer, TemplateSuggestionSerializer
 from .permission import IsAdminOrReadOnly
 
 
@@ -23,7 +24,7 @@ class TemplateTypeViewSet(ModelViewSet):
             queryset = self.queryset.filter(name__iexact=name)
 
         if category:
-            queryset = self.queryset.filter(category__iexact=category)
+            queryset = self.queryset.filter(category__contains=[category])
 
         if id:
             queryset = self.queryset.filter(pk=id)
@@ -41,22 +42,6 @@ class TemplateTypeViewSet(ModelViewSet):
         return Response(serializer.data)
 
 
-class TemplateAIViewSet(ModelViewSet):
-    queryset = TemplateAI.objects.all()
-    serializer_class = TemplateAISerializer
-    permission_classes = [IsAdminOrReadOnly]
-
-    def get_queryset(self):
-        queryset = self.queryset
-        name = self.request.query_params.get('name', None)
-        template_type = self.request.query_params.get('template_type', None)
-        if name:
-            queryset = self.queryset.filter(name__iexact=name)
-        if template_type:
-            queryset = self.queryset.filter(pk=template_type)
-        return queryset
-
-
 class TemplateFeatureViewSet(ModelViewSet):
     queryset = TemplateFeature.objects.all()
     serializer_class = TemplateFeatureSerializer
@@ -68,6 +53,19 @@ class TemplateFeatureViewSet(ModelViewSet):
         name = self.request.query_params.get('name', None)
         if name:
             queryset = self.queryset.filter(name__iexact=name)
+        if id:
+            queryset = self.queryset.filter(pk=id)
+        return queryset
+
+
+class TemplateSuggestionViewSet(ModelViewSet):
+    queryset = TemplateSuggestion.objects.all()
+    serializer_class = TemplateSuggestionSerializer
+    permission_classes = [IsAuthenticated]
+
+    def get_queryset(self):
+        queryset = self.queryset
+        id = self.request.query_params.get('id', None)
         if id:
             queryset = self.queryset.filter(pk=id)
         return queryset